Move typescriptCompiler to the root of the system as it is used by several pieces and is central to the system
