+++ package.json
@@ -9,5 +9,15 @@
   },
   "superpowers": {
     "systemId": "game"
+  },
+  "dependencies": {
+    "combine-source-map": "^0.7.1",
+    "convert-source-map": "^1.1.1",
+    "typescript": "~1.8.2",
+    "uglify-js": "2.7.5"
+  },
+  "devDependencies": {
+    "@types/convert-source-map": "0.0.30",
+    "@types/uglify-js": "^2.6.28"
   }
 }
