+++ plugins/default/typescript/index.d.ts
@@ -1,4 +1,3 @@
-/// <reference path="./node_modules/typescript/lib/typescript.d.ts" />
 /// <reference path="../../common/textEditorWidget/widget.d.ts" />
 /// <reference path="../../../../../SupClient/SupClient.d.ts" />
 /// <reference path="../../../../../SupCore/SupCore.d.ts" />
