+++ typescriptCompiler/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es5",
+    "noImplicitAny": true,
+    "typeRoots": [ "../../../node_modules/@types" ]
+  }
+}
